Issue 276

+++ geonode/base/management/commands/restore.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #########################################################################
 #
-# Copyright (C) 2016 OSGeo
+# Copyright (C) 2018 OSGeo
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -175,6 +175,13 @@ class Command(BaseCommand):
                     gs_data_root = os.path.join(settings.PROJECT_ROOT, '..', gs_data_root)
 
                 try:
+                    helpers.chmod_tree(gs_data_root)
+                except:
+                    print 'Original GeoServer Data Dir "{}" must be writable by the current user. \
+                        Do not forget to copy it first. It will be wiped-out by the Restore procedure!'.format(gs_data_root)
+                    raise
+
+                try:
                     shutil.rmtree(gs_data_root)
                     print 'Cleaned out old GeoServer Data Dir: ' + gs_data_root
                 except:
@@ -225,7 +232,8 @@ class Command(BaseCommand):
     def restore_geoserver_externals(self, config, settings, target_folder):
         """Restore external references from XML files"""
         external_folder = os.path.join(target_folder, helpers.EXTERNAL_ROOT)
-        dir_util.copy_tree(external_folder, '/')
+        if os.path.exists(external_folder):
+            dir_util.copy_tree(external_folder, '/')
 
     def handle(self, **options):
         # ignore_errors = options.get('ignore_errors')
@@ -263,6 +271,37 @@ class Command(BaseCommand):
                 # Extract ZIP Archive to Target Folder
                 target_folder = helpers.unzip_file(backup_file, restore_folder)
 
+                # Write Checks
+                media_root = settings.MEDIA_ROOT
+                media_folder = os.path.join(target_folder, helpers.MEDIA_ROOT)
+                static_root = settings.STATIC_ROOT
+                static_folder = os.path.join(target_folder, helpers.STATIC_ROOT)
+                static_folders = settings.STATICFILES_DIRS
+                static_files_folders = os.path.join(target_folder, helpers.STATICFILES_DIRS)
+                template_folders = settings.TEMPLATE_DIRS
+                template_files_folders = os.path.join(target_folder, helpers.TEMPLATE_DIRS)
+                locale_folders = settings.LOCALE_PATHS
+                locale_files_folders = os.path.join(target_folder, helpers.LOCALE_PATHS)
+
+                try:
+                    print("[Sanity Check] Full Write Access to '{}' ...".format(media_root))
+                    helpers.chmod_tree(media_root)
+                    print("[Sanity Check] Full Write Access to '{}' ...".format(static_root))
+                    helpers.chmod_tree(static_root)
+                    for static_files_folder in static_folders:
+                        print("[Sanity Check] Full Write Access to '{}' ...".format(static_files_folder))
+                        helpers.chmod_tree(static_files_folder)
+                    for template_files_folder in template_folders:
+                        print("[Sanity Check] Full Write Access to '{}' ...".format(template_files_folder))
+                        helpers.chmod_tree(template_files_folder)
+                    for locale_files_folder in locale_folders:
+                        print("[Sanity Check] Full Write Access to '{}' ...".format(locale_files_folder))
+                        helpers.chmod_tree(locale_files_folder)
+                except:
+                    print("...Sanity Checks on Folder failed. Please make sure that the current user has full WRITE access to the above folders (and sub-folders or files).")
+                    print("Reason:")
+                    raise
+
             if not skip_geoserver:
                 self.restore_geoserver_backup(settings, target_folder)
                 self.restore_geoserver_raster_data(config, settings, target_folder)
@@ -275,7 +314,6 @@ class Command(BaseCommand):
             # Prepare Target DB
             try:
                 call_command('migrate', interactive=False, load_initial_data=False)
-                call_command('flush', interactive=False, load_initial_data=False)
 
                 db_name = settings.DATABASES['default']['NAME']
                 db_user = settings.DATABASES['default']['USER']
@@ -283,7 +321,7 @@ class Command(BaseCommand):
                 db_host = settings.DATABASES['default']['HOST']
                 db_passwd = settings.DATABASES['default']['PASSWORD']
 
-                helpers.patch_db(db_name, db_user, db_port, db_host, db_passwd)
+                helpers.patch_db(db_name, db_user, db_port, db_host, db_passwd, settings.MONITORING_ENABLED)
             except:
                 traceback.print_exc()
 
@@ -292,6 +330,22 @@ class Command(BaseCommand):
                 print "Deactivating GeoNode Signals..."
                 designals()
                 print "...done!"
+
+                # Flush DB
+                try:
+                    db_name = settings.DATABASES['default']['NAME']
+                    db_user = settings.DATABASES['default']['USER']
+                    db_port = settings.DATABASES['default']['PORT']
+                    db_host = settings.DATABASES['default']['HOST']
+                    db_passwd = settings.DATABASES['default']['PASSWORD']
+
+                    helpers.flush_db(db_name, db_user, db_port, db_host, db_passwd)
+                except:
+                    try:
+                        call_command('flush', interactive=False, load_initial_data=False)
+                    except:
+                        traceback.print_exc()
+                        raise
 
                 # Restore Fixtures
                 for app_name, dump_name in zip(config.app_names, config.dump_names):
@@ -304,11 +358,9 @@ class Command(BaseCommand):
                         traceback.print_exc()
                         print "WARNING: No valid fixture data found for '"+dump_name+"'."
                         # helpers.load_fixture(app_name, fixture_file)
+                        raise
 
                 # Restore Media Root
-                media_root = settings.MEDIA_ROOT
-                media_folder = os.path.join(target_folder, helpers.MEDIA_ROOT)
-
                 try:
                     shutil.rmtree(media_root)
                 except:
@@ -322,9 +374,6 @@ class Command(BaseCommand):
                 print "Media Files Restored into '"+media_root+"'."
 
                 # Restore Static Root
-                static_root = settings.STATIC_ROOT
-                static_folder = os.path.join(target_folder, helpers.STATIC_ROOT)
-
                 try:
                     shutil.rmtree(static_root)
                 except:
@@ -338,9 +387,6 @@ class Command(BaseCommand):
                 print "Static Root Restored into '"+static_root+"'."
 
                 # Restore Static Root
-                static_root = settings.STATIC_ROOT
-                static_folder = os.path.join(target_folder, helpers.STATIC_ROOT)
-
                 try:
                     shutil.rmtree(static_root)
                 except:
@@ -354,9 +400,6 @@ class Command(BaseCommand):
                 print "Static Root Restored into '"+static_root+"'."
 
                 # Restore Static Folders
-                static_folders = settings.STATICFILES_DIRS
-                static_files_folders = os.path.join(target_folder, helpers.STATICFILES_DIRS)
-
                 for static_files_folder in static_folders:
                     try:
                         shutil.rmtree(static_files_folder)
@@ -373,9 +416,6 @@ class Command(BaseCommand):
                     print "Static Files Restored into '"+static_files_folder+"'."
 
                 # Restore Template Folders
-                template_folders = settings.TEMPLATE_DIRS
-                template_files_folders = os.path.join(target_folder, helpers.TEMPLATE_DIRS)
-
                 for template_files_folder in template_folders:
                     try:
                         shutil.rmtree(template_files_folder)
@@ -392,9 +432,6 @@ class Command(BaseCommand):
                     print "Template Files Restored into '"+template_files_folder+"'."
 
                 # Restore Locale Folders
-                locale_folders = settings.LOCALE_PATHS
-                locale_files_folders = os.path.join(target_folder, helpers.LOCALE_PATHS)
-
                 for locale_files_folder in locale_folders:
                     try:
                         shutil.rmtree(locale_files_folder)
@@ -410,6 +447,8 @@ class Command(BaseCommand):
                     helpers.chmod_tree(locale_files_folder)
                     print "Locale Files Restored into '"+locale_files_folder+"'."
 
+                call_command('collectstatic', interactive=False)
+
                 # Cleanup DB
                 try:
                     db_name = settings.DATABASES['default']['NAME']
@@ -422,8 +461,6 @@ class Command(BaseCommand):
                 except:
                     traceback.print_exc()
 
-                print "Restore finished. Please find restored files and dumps into:"
-
                 return str(target_folder)
 
             finally:
@@ -431,3 +468,9 @@ class Command(BaseCommand):
                 print "Reactivating GeoNode Signals..."
                 resignals()
                 print "...done!"
+
+                call_command('migrate', interactive=False, load_initial_data=False, fake=True)
+
+                print "HINT: If you migrated from another site, do not forget to run the command 'migrate_baseurl' to fix Links"
+                print " e.g.:  DJANGO_SETTINGS_MODULE=my_geonode.settings python manage.py migrate_baseurl --source-address=my-host-dev.geonode.org --target-address=my-host-prod.geonode.org"
+                print "Restore finished. Please find restored files and dumps into:"
